Start of large refactor
Going to move all the ebuilds out of this repo and into a branch of a fork of official gento repo.

Keeping all the profiles in this repo has proved very problematic in terms of profile maintenance; i.e. you need to rebuild the world six times if you adjust a profile.

This approach has its downsides.  That said, I think it will be easier to maintain the ebuilds and the catalyst tooling in a manor more similar to the way the gentoo developers obviously do.